refactor: useKeywordsに切り出す
