+++ package.json
@@ -48,6 +48,7 @@
     "rimraf": "^3.0.2",
     "tailwindcss": "^3.3.2",
     "typescript": "4.9.5",
-    "vite": "^4.3.9"
+    "vite": "^4.3.9",
+    "vite-tsconfig-paths": "^4.3.2"
   }
 }
