+++ vite.config.ts
@@ -2,6 +2,8 @@ import { crx } from '@crxjs/vite-plugin';
 import react from '@vitejs/plugin-react';
 import { join, resolve } from 'path';
 import { defineConfig } from 'vite';
+import tsConfigPaths from "vite-tsconfig-paths"
+
 import manifest from './src/manifest';
 
 export default defineConfig({
@@ -28,10 +30,5 @@ export default defineConfig({
       },
     },
   },
-  plugins: [react(), crx({ manifest })],
-  resolve: {
-    alias: {
-      "~": "./src"
-    }
-  }
+  plugins: [react(), crx({ manifest }), tsConfigPaths()],
 });
